+++ lib/glimmer/swt/tab_folder_proxy.rb
@@ -48,5 +48,6 @@ module Glimmer
         end
       end
     end
+    CTabFolderProxy = TabFolderProxy
   end
 end